Completati i metodi di lettura e inserimento di ricette e ingredienti

+++ app/src/main/java/it/unimib/brewday/database/RicettaDao.java
@@ -26,7 +26,7 @@ public interface RicettaDao {
             "SELECT idRicetta, tipoIngrediente, dosaggioIngrediente " +
                     "FROM IngredienteRicetta WHERE idRicetta = :idRicetta"
     )
-    List<IngredienteRicetta> getIngredientiDellaRicetta(int idRicetta);
+    List<IngredienteRicetta> getIngredientiDellaRicetta(long idRicetta);
 
     @Insert
     long[] insertRicettaIngrediente(List<IngredienteRicetta> listaIngredientiRicetta);
